+++ src/api/v2/events.rs
@@ -139,31 +139,6 @@ pub enum TradeStatus {
   Unknown,
 }
 
-impl TradeStatus {
-  /// Convert a `TradeStatus` into an `order::Status`.
-  pub fn to_order_status(self) -> Option<order::Status> {
-    let status = match self {
-      Self::New => order::Status::New,
-      Self::Replaced => order::Status::Replaced,
-      Self::PartialFill => order::Status::PartiallyFilled,
-      Self::Filled => order::Status::Filled,
-      Self::DoneForDay => order::Status::DoneForDay,
-      Self::Canceled => order::Status::Canceled,
-      Self::Expired => order::Status::Expired,
-      Self::PendingCancel => order::Status::PendingCancel,
-      Self::Stopped => order::Status::Stopped,
-      Self::Rejected => order::Status::Rejected,
-      Self::Suspended => order::Status::Suspended,
-      Self::PendingNew => order::Status::PendingNew,
-      Self::Calculated => order::Status::Calculated,
-      Self::Unknown => order::Status::Unknown,
-      Self::ReplaceRejected | Self::CancelRejected => return None,
-    };
-
-    Some(status)
-  }
-}
-
 
 /// A representation of a trade update that we receive through the
 /// "trade_updates" stream.